feat: add projection head tests for torch
